refactor(mobile-native): split monolithic ListingDetailScreen.kt into cohesive files
Move-only, behavior-preserving split of the 1275-LOC ListingDetailScreen.kt
churn hotspot into 9 cohesive files in the same package/dir. Every moved
Composable body is byte-for-byte identical to the original; the only per-symbol
change is widening `private` -> `internal` on the 14 composables now called
across a file boundary (same androidApp module, no behavior change). The public
entry `ListingDetailScreen` and its signature are unchanged.

Files:
- ListingDetailScreen.kt : entry + ErrorContent
- ListingDetailContent.kt: ListingContent scaffold
- ListingHeroGallery.kt  : HeroGallery, WhitePill
- ListingAgentBar.kt     : StickyAgentBar, SmallVerifiedPill
- ListingHeaderSection.kt: HeaderSection, UppercaseBadge, QuickStatsStrip, StatCard
- ListingDetailTabs.kt   : TabStrip, DescriptionBody, FeaturesChips,
                           BuildingPassportCard, NearbyPreviewCard, PlaceholderSection
- ListingPriceHistory.kt : PriceHistoryCard, PriceHistoryRow
- ListingBottomBar.kt    : BottomActionBar
- ListingDialogs.kt      : InquiryDialog, ShareListingSheet
