+++ mobile-native/androidApp/src/main/java/three/two/bit/ppt/reality/ui/listing/ListingDialogs.kt
@@ -1,0 +1,176 @@
+package three.two.bit.ppt.reality.ui.listing
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import three.two.bit.ppt.reality.R
+import three.two.bit.ppt.reality.listing.*
+
+// ─── Inquiry dialog (preserved) ─────────────────────────────────────────────
+
+@Composable
+internal fun InquiryDialog(
+    listing: ListingDetail,
+    isAuthenticated: Boolean,
+    isSubmitting: Boolean = false,
+    errorMessage: String? = null,
+    onDismiss: () -> Unit,
+    onSubmit: (message: String, name: String?, email: String?, phone: String?) -> Unit,
+) {
+    var message by remember {
+        mutableStateOf(
+            "Hi, I'm interested in ${listing.title}. Please contact me with more information."
+        )
+    }
+    var name by remember { mutableStateOf("") }
+    var email by remember { mutableStateOf("") }
+    var phone by remember { mutableStateOf("") }
+
+    val canSubmit =
+        message.isNotBlank() &&
+            !isSubmitting &&
+            (isAuthenticated || (name.isNotBlank() && email.isNotBlank()))
+
+    AlertDialog(
+        onDismissRequest = { if (!isSubmitting) onDismiss() },
+        title = { Text(stringResource(R.string.inquiry_dialog_title)) },
+        text = {
+            Column {
+                if (!isAuthenticated) {
+                    Card(
+                        colors =
+                            CardDefaults.cardColors(
+                                containerColor = MaterialTheme.colorScheme.tertiaryContainer
+                            )
+                    ) {
+                        Row(
+                            modifier = Modifier.padding(12.dp),
+                            verticalAlignment = Alignment.CenterVertically,
+                        ) {
+                            Icon(
+                                Icons.Default.Info,
+                                contentDescription = null,
+                                tint = MaterialTheme.colorScheme.onTertiaryContainer,
+                            )
+                            Spacer(modifier = Modifier.width(8.dp))
+                            Text(
+                                text = stringResource(R.string.inquiry_sign_in_hint),
+                                style = MaterialTheme.typography.bodySmall,
+                                color = MaterialTheme.colorScheme.onTertiaryContainer,
+                            )
+                        }
+                    }
+                    Spacer(modifier = Modifier.height(12.dp))
+                    OutlinedTextField(
+                        value = name,
+                        onValueChange = { name = it },
+                        label = { Text(stringResource(R.string.label_name_required)) },
+                        modifier = Modifier.fillMaxWidth(),
+                        singleLine = true,
+                        enabled = !isSubmitting,
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
+                    OutlinedTextField(
+                        value = email,
+                        onValueChange = { email = it },
+                        label = { Text(stringResource(R.string.label_email_required)) },
+                        modifier = Modifier.fillMaxWidth(),
+                        singleLine = true,
+                        enabled = !isSubmitting,
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
+                    OutlinedTextField(
+                        value = phone,
+                        onValueChange = { phone = it },
+                        label = { Text(stringResource(R.string.label_phone_optional)) },
+                        modifier = Modifier.fillMaxWidth(),
+                        singleLine = true,
+                        enabled = !isSubmitting,
+                    )
+                    Spacer(modifier = Modifier.height(12.dp))
+                }
+                OutlinedTextField(
+                    value = message,
+                    onValueChange = { message = it },
+                    label = { Text(stringResource(R.string.label_message)) },
+                    modifier = Modifier.fillMaxWidth(),
+                    minLines = 4,
+                    maxLines = 6,
+                    enabled = !isSubmitting,
+                )
+                errorMessage?.let { error ->
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = error,
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.error,
+                    )
+                }
+            }
+        },
+        confirmButton = {
+            Button(
+                onClick = {
+                    onSubmit(
+                        message,
+                        name.takeIf { it.isNotBlank() },
+                        email.takeIf { it.isNotBlank() },
+                        phone.takeIf { it.isNotBlank() },
+                    )
+                },
+                enabled = canSubmit,
+            ) {
+                if (isSubmitting) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(16.dp),
+                        strokeWidth = 2.dp,
+                        color = MaterialTheme.colorScheme.onPrimary,
+                    )
+                    Spacer(modifier = Modifier.width(8.dp))
+                }
+                Text(
+                    if (isSubmitting) stringResource(R.string.inquiry_sending)
+                    else stringResource(R.string.inquiry_send)
+                )
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss, enabled = !isSubmitting) {
+                Text(stringResource(R.string.cancel))
+            }
+        },
+    )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun ShareListingSheet(listing: ListingDetail, onDismiss: () -> Unit) {
+    ModalBottomSheet(onDismissRequest = onDismiss) {
+        Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+            Text(
+                text = stringResource(R.string.share_property),
+                style = MaterialTheme.typography.titleLarge,
+                fontWeight = FontWeight.Bold,
+            )
+            Spacer(modifier = Modifier.height(16.dp))
+            ListItem(
+                headlineContent = { Text(stringResource(R.string.action_share_link)) },
+                leadingContent = { Icon(Icons.Default.Link, contentDescription = null) },
+                modifier = Modifier.fillMaxWidth(),
+            )
+            ListItem(
+                headlineContent = { Text(stringResource(R.string.share_via)) },
+                leadingContent = { Icon(Icons.Default.Share, contentDescription = null) },
+                modifier = Modifier.fillMaxWidth(),
+            )
+            Spacer(modifier = Modifier.height(32.dp))
+        }
+    }
+}
